refactor: reorganize schemas and services
