+++ src/database/schemas/article.schema.ts
