+++ src/database/schemas/content.schema.ts
