+++ src/database/schemas/latest_news.schema.ts
