+++ src/database/version.service.spec.ts
@@ -1,12 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { NewsService } from './news.service';
+import { VersionService } from './version.service';
 import { getModelToken } from '@nestjs/mongoose';
-import { News, NewsSchema } from '../../schemas/news.schema';
+import { Version, VersionSchema } from './schemas/version.schema';
 import * as mongoose from 'mongoose';
 
-describe('NewsService', () => {
-  let service: NewsService;
-  let model: mongoose.Model<News>;
+describe('VersionService', () => {
+  let service: VersionService;
+  let model: mongoose.Model<Version>;
 
   beforeAll(async () => {
     await mongoose.connect(global.__MONGO_URI__);
@@ -19,16 +19,16 @@ describe('NewsService', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
-        NewsService,
+        VersionService,
         {
-          provide: getModelToken(News.name),
-          useValue: mongoose.model(News.name, NewsSchema),
+          provide: getModelToken(Version.name),
+          useValue: mongoose.model(Version.name, VersionSchema),
         },
       ],
     }).compile();
 
-    service = module.get<NewsService>(NewsService);
-    model = module.get<mongoose.Model<News>>(getModelToken(News.name));
+    service = module.get<VersionService>(VersionService);
+    model = module.get<mongoose.Model<Version>>(getModelToken(Version.name));
   });
 
   afterEach(async () => {
@@ -41,51 +41,49 @@ describe('NewsService', () => {
     expect(model).toBeDefined();
   });
 
-  describe('findOneById', () => {
-    it('should find one news by id', async () => {
+  describe('findOneByArticleId', () => {
+    it('should find one version by article id', async () => {
       const targetId = new mongoose.Types.ObjectId(1);
 
-      const spyFindById = jest.spyOn(model, 'findById');
-      await service.findOneById(targetId);
+      const spyFindOne = jest.spyOn(model, 'findOne');
+      await service.findOneByArticleId(targetId);
 
-      expect(spyFindById).toHaveBeenCalledTimes(1);
-      expect(spyFindById).toHaveBeenCalledWith(targetId);
+      expect(spyFindOne).toHaveBeenCalledTimes(1);
+      expect(spyFindOne).toHaveBeenCalledWith({ articleId: targetId });
     });
   });
 
   describe('createOne', () => {
-    it('should create one news', async () => {
-      const mockNews: News = {
-        newsId: new mongoose.Types.ObjectId(1),
-        title: 'test',
+    it('should create one version', async () => {
+      const mockVersion: Version = {
+        articleId: new mongoose.Types.ObjectId(1),
       };
 
       const spyCreate = jest.spyOn(model, 'create');
-      await service.createOne(mockNews);
+      await service.createOne(mockVersion);
 
       expect(spyCreate).toHaveBeenCalledTimes(1);
-      expect(spyCreate).toHaveBeenCalledWith(mockNews);
+      expect(spyCreate).toHaveBeenCalledWith(mockVersion);
     });
   });
 
   describe('updateOneById', () => {
-    it('should update one news', async () => {
+    it('should update one version', async () => {
       const targetId = new mongoose.Types.ObjectId(1);
-      const mockNews: News = {
-        newsId: new mongoose.Types.ObjectId(1),
-        title: 'test',
+      const mockVersion: Version = {
+        articleId: new mongoose.Types.ObjectId(1),
       };
 
       const spyUpdateOne = jest.spyOn(model, 'updateOne');
-      await service.updateOneById(targetId, mockNews);
+      await service.updateOneById(targetId, mockVersion);
 
       expect(spyUpdateOne).toHaveBeenCalledTimes(1);
-      expect(spyUpdateOne).toHaveBeenCalledWith({ _id: targetId }, mockNews);
+      expect(spyUpdateOne).toHaveBeenCalledWith({ _id: targetId }, mockVersion);
     });
   });
 
   describe('deleteOneById', () => {
-    it('should delete one news', async () => {
+    it('should delete one version', async () => {
       const targetId = new mongoose.Types.ObjectId(1);
 
       const spyDeleteOne = jest.spyOn(model, 'deleteOne');
